requirements.txt
tensorflow>=2.12.0
keras>=2.12.0
numpy>=1.21.0
Flask>=2.0.1
Flask-Login>=0.6.3
Pillow>=8.3.1
werkzeug>=2.0.1 
gunicorn>=20.1.0
requests>=2.31.0
fastapi>=0.110.0
uvicorn[standard]>=0.23.0